refactor(cognitive_complexity): extract path orchestration into runner module
